remove unwanted fields,add camera option to image upload fields

+++ lib/models/total_loan_count.model.dart
@@ -22,7 +22,7 @@ class TotalLoanCountMessage {
   int? totalEmis;
   double? collectionInHand;
   int? todaysCollection;
-  int? monthlyCollection;
+  double? monthlyCollection;
 
   TotalLoanCountMessage({
     this.totalLoans,
